test: keep the stream socket benchmark's pacing deadline signed
kPacketInterval * i with a size_t i gives the deadline an unsigned
duration; once the producer falls behind, libc++'s sleep_until turns the
negative remaining time into a near-infinite sleep and the benchmark
hangs. Multiply by a signed index.

+++ tests/zm_stream_socket_bench.cpp
@@ -212,7 +212,10 @@ Result RunScenario(size_t consumer_count) {
   auto start = std::chrono::steady_clock::now();
   for (size_t i = 0; i < kPackets; ++i) {
     bool key = (i % kGop) == 0;
-    std::this_thread::sleep_until(start + kPacketInterval * i);
+    // Signed multiplier: a size_t would give the deadline an unsigned rep, and
+    // libc++'s sleep_until underflows to a near-infinite sleep once the
+    // deadline is already in the past.
+    std::this_thread::sleep_until(start + kPacketInterval * static_cast<int64_t>(i));
     double cpu0 = cpu_seconds(CLOCK_THREAD_CPUTIME_ID);
     auto t0 = std::chrono::steady_clock::now();
     server.SendMedia(key ? keyframe.get() : delta.get(), StreamId::Video, key,
